Several Fixes and additions
Changed the way the equation is built, enhancing checking capabilities
Added buylist saving.
Buylists are now saved in the same directory they came from with the suffix "_buylist" added.
Script now checks if the list was already run, to allow for recalculations (perhaps changing the banned_store list)
Script now checks and respects a banlist. Add stores to the banned_stores.txt to ignore them when making the buylist.